修复
统一项目参数的位置
清除无用的方法
修复没生效的mixin
删除不生效的代码

+++ src/main/java/com/smd/gctcore/common/mixin/moretcon/MixinRing.java
@@ -26,7 +26,7 @@ public abstract class MixinRing {
                 }
             }
             if (count == 0) return false;
-            return player.getEntityWorld().getWorldTime() % (count * count) == 0;
+            return player.getEntityWorld().getWorldTime() % ((long) count * count) == 0;
         }
         return false;
     }
